Move use cases operating on flow repository to own module
