+++ OTAnalytics/application/application.py
@@ -25,17 +25,12 @@ from OTAnalytics.application.state import (
 from OTAnalytics.application.use_cases.config import SaveConfiguration
 from OTAnalytics.application.use_cases.event_repository import ClearEventRepository
 from OTAnalytics.application.use_cases.export_events import EventListExporter
+from OTAnalytics.application.use_cases.flow_repository import AddFlow
 from OTAnalytics.application.use_cases.section_repository import AddSection
 from OTAnalytics.application.use_cases.update_project import ProjectUpdater
 from OTAnalytics.domain.date import DateRange
 from OTAnalytics.domain.filter import FilterElement, FilterElementSettingRestorer
-from OTAnalytics.domain.flow import (
-    Flow,
-    FlowChangedObserver,
-    FlowId,
-    FlowListObserver,
-    FlowRepository,
-)
+from OTAnalytics.domain.flow import Flow, FlowChangedObserver, FlowId, FlowListObserver
 from OTAnalytics.domain.geometry import RelativeOffsetCoordinate
 from OTAnalytics.domain.section import (
     Section,
@@ -56,41 +51,12 @@ class CancelAddFlow(Exception):
     pass
 
 
-class FlowAlreadyExists(Exception):
-    pass
-
-
 class MultipleSectionsSelected(Exception):
     pass
 
 
 class MultipleFlowsSelected(Exception):
     pass
-
-
-class AddFlow:
-    """
-    Add a single flow to the repository.
-    """
-
-    def __init__(self, flow_repository: FlowRepository) -> None:
-        self._flow_repository = flow_repository
-
-    def add(self, flow: Flow) -> None:
-        if not self.is_flow_name_valid(flow.name):
-            raise FlowAlreadyExists(
-                f"A flow with the name {flow.name} already exists. "
-                "Choose another name."
-            )
-        self._flow_repository.add(flow)
-
-    def is_flow_name_valid(self, flow_name: str) -> bool:
-        if not flow_name:
-            return False
-        return all(
-            stored_flow.name != flow_name
-            for stored_flow in self._flow_repository.get_all()
-        )
 
 
 class IntersectTracksWithSections(ABC):
