+++ tests/OTAnalytics/application/use_cases/test_flow_repository.py
@@ -1,0 +1,35 @@
+from unittest.mock import Mock, call
+
+import pytest
+
+from OTAnalytics.application.use_cases.flow_repository import AddFlow, FlowAlreadyExists
+from OTAnalytics.domain.flow import Flow, FlowRepository
+
+
+class TestAddFlow:
+    def test_add_flow_with_different_names(self) -> None:
+        some_flow = Mock(spec=Flow)
+        some_flow.name = "some"
+        other_flow = Mock(spec=Flow)
+        other_flow.name = "other"
+        flow_repository = Mock(spec=FlowRepository)
+        flow_repository.get_all.return_value = [some_flow]
+        use_case = AddFlow(flow_repository)
+
+        use_case.add(other_flow)
+
+        assert flow_repository.add.call_args_list == [
+            call(other_flow),
+        ]
+
+    def test_add_flow_with_same_names(self) -> None:
+        some_flow = Mock(spec=Flow)
+        some_flow.name = "some"
+        other_flow = Mock(spec=Flow)
+        other_flow.name = "some"
+        flow_repository = Mock(spec=FlowRepository)
+        flow_repository.get_all.return_value = [some_flow]
+        use_case = AddFlow(flow_repository)
+
+        with pytest.raises(FlowAlreadyExists):
+            use_case.add(other_flow)
